# Optimize pattern-inside's ending with `...`