--- a/semgrep-core/src/parsing/Parse_mini_rule.ml
+++ b/semgrep-core/src/parsing/Parse_mini_rule.ml
@@ -106,6 +106,7 @@ let parse file =
                          {
                            R.id;
                            pattern;
+                           inside = false;
                            message;
                            languages;
                            severity;