This probably won't have a huge perf impact overall due to parsing being
the main bottle neck. However it does help perf and it's a fairly simple
optimization.

test plan:

    % make test

    % semgrep-core -lang js -config ~/semgrep-rules/typescript/react/security/audit/react-css-injection.yaml ~/semgrep/perf/bench/big-js/input/big-js
      #^ Now takes about ~19s (vs ~27s, or ~1.4x faster)
      # -profile reports matching time is ~2s (vs ~10s, or ~5x faster)

    % semgrep-core -profile -lang js \
         -config semgrep-rules/typescript/react/security/audit/react-props-injection.yaml \
         semgrep/parsing-stats/lang/javascript/tmp/mui-org-material-ui/test/bundling/fixtures/next-webpack5
      #^ Now takes about ~0.1s (vs ~1.5s, or ~12x faster)

    % semgrep -j1 --time --max-target-bytes=250000 --config semgrep-rules/typescript/react/security/audit/react-props-injection.yaml semgrep/parsing-stats/lang/javascript/tmp/*
      #^ Match time dropped from ~20s to ~10 (2x faster), but overall
      # there is not much of a speed up since parsing takes ~80% of
      # the time.



PR checklist:
- [x] documentation is up to date
- [x] changelog is up to date